# packages/founder/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
    "/api/v1/user-context": {
      "get": {
        "operationId": "UserContextController_getUserContext",
        "summary": "Get aggregated user context",
        "description": "Aggregates data from multiple external APIs (location, weather, cat facts). Supports different concurrency strategies.",
        "parameters": [
          {
            "name": "strategy",
            "required": false,
            "in": "query",
            "description": "Aggregation strategy to use (default: promise-allsettled)",
            "schema": {
              "enum": [
                "promise-allsettled",
                "rxjs",
                "async-sequential"
              ],
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "User context successfully aggregated",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserContextResponseDto"
                }
              }
            }
          },
          "500": {
            "description": "Internal server error"
          }
        },
        "tags": [
          "user-context"
        ]
      }
    },
    "/health": {
      "get": {
        "operationId": "HealthController_getHealth",
        "summary": "Health check endpoint",
        "parameters": [],
        "responses": {
          "200": {
            "description": "Service is healthy",
            "content": {
              "application/json": {
                "schema": {
                  "properties": {
                    "status": {
                      "type": "string",
                      "example": "healthy"
                    },
                    "timestamp": {
                      "type": "string",
                      "example": "2024-01-26T10:30:00.000Z"
                    }
                  }
                }
              }
            }
          }
        },
        "tags": [
          "health"
        ]
      }
    }
  },
  "info": {
    "title": "Fake Stripe - Founder API",
    "description": "User context aggregator service combining location, weather, and entertainment data using hexagonal architecture with multiple concurrency strategies",
    "version": "1.0",
    "contact": {}
  },
  "tags": [
    {
      "name": "user-context",
      "description": "User context aggregation endpoints"
    },
    {
      "name": "health",
      "description": "Health check endpoints"
    }
  ],
  "servers": [],
  "components": {
    "schemas": {
      "LocationDto": {
        "type": "object",
        "properties": {
          "ip": {
            "type": "string",
            "example": "203.0.113.42",
            "description": "IP address"
          },
          "city": {
            "type": "string",
            "example": "San Francisco",
            "description": "City name"
          },
          "region": {
            "type": "string",
            "example": "California",
            "description": "Region/state name"
          },
          "country": {
            "type": "string",
            "example": "US",
            "description": "Country code"
          },
          "latitude": {
            "type": "number",
            "example": 37.7749,
            "description": "Latitude coordinate"
          },
          "longitude": {
            "type": "number",
            "example": -122.4194,
            "description": "Longitude coordinate"
          },
          "timezone": {
            "type": "string",
            "example": "America/Los_Angeles",
            "description": "Timezone identifier"
          }
        },
        "required": [
          "ip",
          "city",
          "region",
          "country",
          "latitude",
          "longitude",
          "timezone"
        ]
      },
      "WeatherDto": {
        "type": "object",
        "properties": {
          "description": {
            "type": "string",
            "example": "Clear sky",
            "description": "Weather description"
          },
          "temperature": {
            "type": "number",
            "example": 18.5,
            "description": "Temperature in Celsius"
          },
          "feelsLike": {
            "type": "number",
            "example": 17.2,
            "description": "Feels like temperature in Celsius"
          },
          "humidity": {
            "type": "number",
            "example": 65,
            "description": "Humidity percentage"
          },
          "windSpeed": {
            "type": "number",
            "example": 3.5,
            "description": "Wind speed in m/s"
          }
        },
        "required": [
          "description",
          "temperature",
          "feelsLike",
          "humidity",
          "windSpeed"
        ]
      },
      "EntertainmentDto": {
        "type": "object",
        "properties": {
          "catFact": {
            "type": "string",
            "example": "Cats sleep 70% of their lives.",
            "description": "Random cat fact"
          },
          "source": {
            "type": "string",
            "example": "catfact.ninja",
            "description": "Source of the fact"
          }
        },
        "required": [
          "catFact",
          "source"
        ]
      },
      "UserContextResponseDto": {
        "type": "object",
        "properties": {
          "location": {
            "description": "User location data (null if failed)",
            "allOf": [
              {
                "$ref": "#/components/schemas/LocationDto"
              }
            ]
          },
          "weather": {
            "description": "Weather data (null if failed)",
            "allOf": [
              {
                "$ref": "#/components/schemas/WeatherDto"
              }
            ]
          },
          "entertainment": {
            "description": "Entertainment data (null if failed)",
            "allOf": [
              {
                "$ref": "#/components/schemas/EntertainmentDto"
              }
            ]
          },
          "aggregatedAt": {
            "format": "date-time",
            "type": "string",
            "example": "2024-01-26T10:30:00.000Z",
            "description": "Timestamp of aggregation"
          },
          "processingTimeMs": {
            "type": "number",
            "example": 450,
            "description": "Processing time in milliseconds"
          },
          "strategyUsed": {
            "type": "string",
            "example": "promise-allsettled",
            "description": "Strategy used for aggregation",
            "enum": [
              "promise-allsettled",
              "rxjs",
              "async-sequential"
            ]
          },
          "errors": {
            "example": [
              "Weather fetch failed: timeout"
            ],
            "description": "Array of errors if any API calls failed",
            "type": "array",
            "items": {
              "type": "string"
            }
          }
        },
        "required": [
          "aggregatedAt",
          "processingTimeMs",
          "strategyUsed"
        ]
      }
    }
  }
}